introduce the NTERM-CASE form to the example risc-pda
There was previously no way to dispatch on a non-terminal. The code was,
incorrectly, using state-case to dispatch on non-terminals. I could have
solved this by expanding every (non-term . insns) into a set of clauses
with a clause for each rule associated with the non-terminal. However, that
doesn't seem like the right solution because it might be impossible for some
rules to reduce to the given state. Rather than waste space, I added a
nterm-case. I do not believe this violates any of our assumptions about the
PDA.