Modify the compiler to produce code that type checks for the stream type system
I had to add labeled code blocks for both the eos version of each state and each
rule. This necessitated a couple new procedures for compiling an eos state. This
change demanded a change to the `compile-action' procedure.

I also made a small change, which supports eos rule blocks, to compile-rule.