Introduce a series of syntax-classes describing how to parse PDAs
Each syntax-class describes a piece of the PDA. The interface to this series of
syntax-classes are typed-pda-clauses-stx and untyped-pda-clauses-stx which match
a series of typed and an untyped pda clauses, respectively. The composing
clauses are states, rules, eos, start, tokens, and comment. The syntax-classes
for states are composed of syntax-classes for the four actions: shift, reduce,
goto, and accept.

The top-level macro is stored in parse-pda-macro.rkt, and is a simple wrapper
around parse-pda, which is stored in parse-pda.rkt. The parse-pda procedure is a
syntax-parse which matches one of the previously mentioned pda-clauses syntax
classes. Segregating the logic from the macro allows for normal rackunit testing
of the logic. Furthermore, experimenting with the logic code when it is used at
phase 0 is much easier because phase 0 functions are not required to return
syntax.

The parse-pda procedure produces an intermediate representation based on the PDA
ADT from pda-data.rkt. This IR will eventually be passed to the PDA compiler and
the PDA-RISC compiler, thus completing the production of Racket code.

  A Note Regarding Syntax-Case

In syntax-parse, literals must be defined. This ensures that literals are not
used outside their intended forms. Ryan explains it here:
  http://macrologist.blogspot.com/2011/09/syntax-parse-and-literals.html
However, I don't believe his explanation is particularly insightful.