In preperation for the parsing Syntax instead of SExps, add Syntax to PDA ADT
The PDA ADT now expects Syntax for state names, rule names, and tokens. These
all must be passed as Syntax to ensure hygiene later on. Non-terminals are not
important because the cannot escape the PDA form. Tokens, however, must match up
to the tokens the user specifies because the macro will use a user supplied
token-case form to deal with the input-stream.