Fix handling of non-* block insns
Note that in the prior circumstance, a term's successors was always the next
item in the term*-seq. This is incorrect because block's introduce nested
structure to term*-seqs.

Now, `set-preds/term!' checks for for block insns and enters the mutually
recursive procedure `set-preds/term-seq!' which properly handles this nested
structure.

This nested structure also complicates the predecessor relationship in an
insn*-seq. If an insn*-seq has the form

  (list ... A B ...)

then B's predecessor is A, iff A is not a (block ...). If A is, in fact, a

  (block I ... LAST)

then B's predecessor is LAST.