Fix bug in setting predecessors of insn block sub-insns
Previously, we set the predecessor of the first insn of a block to the
predecessor of the block. This creates a discrepancy because the forward chain
is longer than the backwards chain (which skips the block).

This change sets the first insn of a block to the block itself.

This bug did not affect block*s.